Hubi_Telegram/requirements.txt
python-telegram-bot
VaderSentiment
openai
regex
nltk
